requirements.txt
Flask               >= 0.10, < 0.11
Flask-SQLAlchemy    >= 2.0, < 3.0
Flask-Migrate       >= 1.6, < 1.7
Flask-Security      == 1.7.4
Flask-Login         == 0.2.11
Flask-WTF           >= 0.12, < 0.13
Flask-Bootstrap     >= 3.3, < 3.4
Flask-Script        >= 2.0, < 2.1
Flask-Mail          >= 0.9, < 0.10
waitress            >= 0.8, < 0.9
psycopg2            >= 2.6, < 2.7
bcrypt              >= 2.0, < 2.1
cachetools          >= 1.1, < 1.2
PyYAML              >= 3.11, < 3.12
docker-py           >= 1.7, < 1.8
docker-compose      >= 1.6, < 1.7
requests            >= 2.7, < 2.8
dumb-init           >= 1.0, < 1.1
